Reorganizing the repository to include 3 submodules (gathering training data, training/testing the network, and the actual binary)
